requirements.txt
streamlit>=1.40.0
openai>=1.12.0
pydantic>=2.7.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
